+++ SdkTests/Native/Functions/Functions.cpp
@@ -179,6 +179,50 @@ DECL(MyEnum) FirstEnumElement(MyEnum test[])
     return test[0];
 }
 
+DECL(int) ArrayRelationSum(int length, SimpleStruct elements[])
+{
+    return Sum(length, elements);
+}
+
+DECL(void) ArrayRelationOutInitBoolArray(bool out[], int length)
+{
+    for(size_t i = 0; i < length; i++)
+    {
+        out[i] = true;
+    }
+}
+
+DECL(void) ArrayRelationOutGetInterfacesWithRelation(int length, Interface** array)
+{
+    GetInterfaces(length, array);
+}
+
+DECL(void) ArrayRelationInInterfaceArray(int length, Interface* array[])
+{
+    for(size_t i = 0; i < length; i++)
+    {
+        array[i]->Method();
+    }
+}
+
+DECL(int) ArrayRelationSumStructWithMarshal(int length, StructWithMarshal array[])
+{
+    int sum = 0;
+    for(size_t i = 0; i < length; i++)
+    {
+        for(size_t j = 0; j < 3; j++)
+        {
+            sum += array[i].i[j];
+        }
+    }
+    return sum;
+}
+
+DECL(bool) VerifyReservedParam(int reserved)
+{
+    return reserved == 42;
+}
+
 DECL(StructAsClassWrapper) GetWrapper()
 {
     return {{1}};