+++ SharpGen/Generator/Marshallers/ArrayOfInterfaceMarshaller.cs
@@ -109,7 +109,7 @@ namespace SharpGen.Generator.Marshallers
 
         public IEnumerable<StatementSyntax> GenerateNativeToManagedExtendedProlog(CsMarshalCallableBase csElement)
         {
-            yield return NotImplemented("Array of interfaces");
+            yield return GenerateArrayNativeToManagedExtendedProlog(csElement);
         }
 
         public FixedStatementSyntax GeneratePin(CsParameter csElement)