+++ SharpGen/Generator/Marshallers/MarshallerBase.cs
@@ -4,6 +4,7 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using SharpGen.Model;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
 namespace SharpGen.Generator.Marshallers
@@ -319,6 +320,21 @@ namespace SharpGen.Generator.Marshallers
                                     Literal(message)))))));
         }
 
+        protected StatementSyntax NotSupported(string message)
+        {
+            return ThrowStatement(
+                ObjectCreationExpression(
+                    ParseTypeName("System.NotSupportedException"))
+                .WithArgumentList(
+                    ArgumentList(
+                        message == null ? default
+                        : SingletonSeparatedList(
+                            Argument(
+                                LiteralExpression(
+                                    SyntaxKind.StringLiteralExpression,
+                                    Literal(message)))))));
+        }
+
         protected ArgumentSyntax GenerateManagedValueTypeArgument(CsParameter csElement)
         {
             var arg = Argument(IdentifierName(csElement.Name));
@@ -370,6 +386,23 @@ namespace SharpGen.Generator.Marshallers
             return param;
         }
 
+        protected StatementSyntax GenerateArrayNativeToManagedExtendedProlog(CsMarshalCallableBase csElement)
+        {
+            var callable = (CsCallable)csElement.Parent;
+            var lengthParam = callable.Parameters
+                .FirstOrDefault(param
+                    => param.Relation is LengthRelation relation
+                        && relation.RelatedMarshallableName == csElement.CppElementName);
+            if (lengthParam != null)
+            {
+                var marshaller = new LengthRelationMarshaller(globalNamespace);
+                return marshaller.GenerateNativeToManaged(csElement, lengthParam);
+            }
+            else
+            {
+                return NotSupported("Cannot marshal a native array to a managed array when length is not specified");
+            }
+        }
         protected StatementSyntax GenerateGCKeepAlive(CsMarshalBase csElement)
         {
             return ExpressionStatement(