+++ SharpGen/Generator/ReverseCallablePrologCodeGenerator.cs
@@ -42,7 +42,7 @@ namespace SharpGen.Generator
             {
                 var publicParameter = csElement.Parameters[i];
                 var nativeParameter = IdentifierName($"param{i}");
-                var prologBuilder = publicParameter.PassedByNativeReference
+                var prologBuilder = publicParameter.PassedByNativeReference && !publicParameter.IsArray
                     ? (Func<CsMarshalCallableBase, ExpressionSyntax, IEnumerable<StatementSyntax>>)GenerateNativeByRefProlog
                     : GenerateProlog;
                 foreach (var statement in prologBuilder(publicParameter, nativeParameter))
@@ -57,13 +57,18 @@ namespace SharpGen.Generator
             ExpressionSyntax nativeParameter)
         {
             var marshaller = generators.Marshalling.GetMarshaller(publicElement);
+            var publicType = ParseTypeName(publicElement.PublicType.QualifiedName);
+            if (publicElement.IsArray)
+            {
+                publicType = ArrayType(publicType, SingletonList(ArrayRankSpecifier()));
+            }
             yield return LocalDeclarationStatement(
-                VariableDeclaration(ParseTypeName(publicElement.PublicType.QualifiedName))
+                VariableDeclaration(publicType)
                 .AddVariables(
                     VariableDeclarator(Identifier(publicElement.Name))
                         .WithInitializer(
                             EqualsValueClause(
-                                DefaultExpression(ParseTypeName(publicElement.PublicType.QualifiedName))))));
+                                DefaultExpression(publicType)))));
 
             if (marshaller.GeneratesMarshalVariable(publicElement))
             {